requirements.txt
appdirs==1.4.2
Django==1.10.6
django-filter==1.0.1
djangorestframework==3.5.4
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
pkg-resources==0.0.0
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
